Update loader/adaptative.py with loader/adaptative_new.py
